Refactor testing library and add missing breaks
