+++ opsdroid/testing/fixtures.py
@@ -2,16 +2,27 @@
 import pytest
 
 from opsdroid.core import OpsDroid
+from .external_api import ExternalAPIMockServer
 
-__all__ = ["opsdroid"]
+__all__ = ["opsdroid", "mock_api"]
 
 
 @pytest.fixture
-def opsdroid():
+def opsdroid() -> OpsDroid:
     """Fixture with a plain instance of opsdroid.
 
-    Will yield an instance of opsdroid which hasn't been loaded.
+    Will yield an instance of :class:`opsdroid.core.OpsDroid` which hasn't been loaded.
 
     """
     with OpsDroid(config={}) as opsdroid:
         yield opsdroid
+
+
+@pytest.fixture
+async def mock_api() -> ExternalAPIMockServer:
+    """Fixture for making a mock API server.
+
+    Will give an instance of :class:`opsdroid.testing.ExternalAPIMockServer`.
+
+    """
+    return ExternalAPIMockServer()
